+++ lib/pages/settings_page.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:url_launcher/url_launcher.dart';
+import '../constants.dart';
 import '../l10n/app_localizations.dart';
 import '../models/server_config.dart';
 import '../app.dart' show presetColors, appVersion;
@@ -30,7 +31,7 @@ class SettingsPage extends StatefulWidget {
 class _SettingsPageState extends State<SettingsPage> {
   AppLocalizations get l10n => AppLocalizations(widget.config.language);
 
-  static const _ch = MethodChannel('cc.merr.inout/native');
+  static const _ch = MethodChannel(kMethodChannel);
   // Cache permission state across page rebuilds
   bool? _cachedStorageGranted;
 
@@ -87,7 +88,7 @@ class _SettingsPageState extends State<SettingsPage> {
             child: Column(
               children: [
                 Text(
-                  'INOUT',
+                  'DUFSHUB',
                   style: TextStyle(
                     fontFamily: 'PressStart2P',
                     fontSize: 24,
@@ -112,11 +113,11 @@ class _SettingsPageState extends State<SettingsPage> {
                 const SizedBox(height: 8),
                 GestureDetector(
                   onTap: () => launchUrl(
-                    Uri.parse('https://github.com/zocs/inout'),
+                    Uri.parse('https://github.com/zocs/dufshub'),
                     mode: LaunchMode.externalApplication,
                   ),
                   child: Text(
-                    'github.com/zocs/inout',
+                    'github.com/zocs/dufshub',
                     style: Theme.of(context).textTheme.bodySmall?.copyWith(
                       color: Theme.of(context).colorScheme.primary,
                       fontWeight: FontWeight.w500,
@@ -145,7 +146,7 @@ class _SettingsPageState extends State<SettingsPage> {
                       ),
                       onPressed: () => showLicensePage(
                         context: context,
-                        applicationName: 'inout',
+                        applicationName: 'DufsHub',
                         applicationVersion: 'v$appVersion',
                         applicationLegalese:
                             'Copyright (c) 2026 zocs\nMIT License',